images2pdf/requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
tqdm>=4.65.0
# Optional for OCR
pytesseract>=0.3.10
# Optional for HEIC support
pillow-heif>=0.13.0
